[Wony][SDISel] Add i16 load instruction (ldr16) The new LDR16 instruction which is a load from an address described by a base pointer (in register) and an offset (4-bit unsigned immediate value)
We implement this as a complex pattern.
A complex pattern allows to write custom C++ matchers, meaning that we can
decide to apply a pattern based on some C++ code instead of relying purely
on the TableGen types and nodes.

Additionally, a complex pattern allows to take an input node and split
it in different operands.

This type of matches allows for instance to fold an address computation
directly in the load. For instance, if the input address is defined by
base + offset, we can fold this addition directly in the load.

Signed-off-by: ZakyHermawan <zaky.hermawan9615@gmail.com>