[Wony] Implement call instruction
The CALL instruction consists of just a jump to an address
represented by an 11-bit immediate value.

Update:
- Our AsmPrinter to teach it how to handle symbols
- Our MCCodeEmitter to teach how to add a fix-up for the call target

A fix-up is an indication left by the assembler to the linker that tells
the linker that it needs to patch the final binary with something.
In this specific case, the fix-up would tell the liker to replace the 11-bit
of the immediate value with the actual address (PC-relative) of the related
symbol.
The fix-up is currently not emitted because we didn't connect a target
specific MCAsmBackend yet.

Signed-off-by: ZakyHermawan <zaky.hermawan9615@gmail.com>