[Wony][GISel] Setup GlobalISel
This commit implements a skeleton of the GlobalISel pipeline for our target.
The GISel pipeline consists in four mandatory phases:
1. IRTranslator, which requires a XXXCallLowering instance
2. Legalizer, which requires a XXXLegalizerInfo instance
3. RegisterBankSelect, which requires a XXXRegisterBankInfo
4. InstructionSelect, which requires a XXXInstructionSelector

In this commit we only provide the WonyCallLowering and test the pipeline
until after IRTranslator.

The implementation of WonyCallLowering may look complex, but this is
just a reimplementation of everything that has been done in SDISel so far.

Signed-off-by: ZakyHermawan <zaky.hermawan9615@gmail.com>