[Wony][IRTranslator] Support structure lowering
This commit teaches the GlobalISel Wony backend how to perform sret demoting.
SRet demoting consists in returning a structure through the stack instead of
breaking it up into potentially several registers.

The caller is responsible for providing the stack slot as an input argument to
the callee and the callee is responsible for writing the structure to be
returned in that stack slot.

This is called demoting because the structure is demoted from a fast storage
(register) to a slow storage (memory).

Putting things together SRet demoting means going from:
void caller() {
  myStruct = callee()
}

To
void caller() {
  myStruct;
  callee(&myStruct);
}

Signed-off-by: ZakyHermawan <zaky.hermawan9615@gmail.com>